requirements-develop.txt
certifi==2021.10.8
charset-normalizer==2.0.9
coverage==6.2
idna==3.3
PyYAML==6.0
requests==2.26.0
requests-mock==1.9.3
six==1.16.0
urllib3==1.26.7
